Made qx add a super command like qx contrib
