+++ commands/add_commands/class.js
@@ -1,0 +1,19 @@
+/* ************************************************************************
+
+   qooxdoo - the new era of web development
+
+   http://qooxdoo.org
+
+   Copyright:
+     2017 Christian Boulanger
+
+   License:
+     MIT: https://opensource.org/licenses/MIT
+     See the LICENSE file in the project's top-level directory for details.
+
+   Authors:
+     * Christian Boulanger (info@bibliograph.org, @cboulanger)
+
+************************************************************************ */
+require("../../lib/qxcli/commands/add/Class");
+module.exports = qxcli.commands.add.Class.getYargsCommand();
