+++ lib/qxcli/commands/Add.js
@@ -5,7 +5,7 @@
    http://qooxdoo.org
 
    Copyright:
-     2017 Christian Boulanger and others
+     2017 Christian Boulanger
 
    License:
      MIT: https://opensource.org/licenses/MIT
@@ -16,184 +16,35 @@
 
 ************************************************************************ */
 
-/*global qx qxcli*/
-
-const qx = require("qooxdoo");
-const fs = require("fs");
-const process = require('process');
-const path = require('upath');
-const jsonlint = require("jsonlint");
-const {promisify} = require('util');
-const JsonToAst = require("json-to-ast");
-const readFile = promisify(fs.readFile);
-
 require("./Command");
 
+const qx = require("qooxdoo");
+
 /**
- * Add a new class file to the current project, based on a template.
- *
- * Syntax: `qx add template_name class_name [--extend=extended_class] [--import] [--from-contrib=contrib_name]`
- * The path to the template file will be calculated as follows:
- * 1. transform template_name to ${template_name}.tmpl.js
- * 2. check if this file exists in the "templates/class" folder 
- *    a. of a contrib, if the --from-contrib paramater wass passed
- *    b. otherwise, of the current project
- *    c. finally, of the CLI library
- *
- * If the --import flag is set, copy the file to the templates/class folder of the current project.
- * 
- * If you place a file named `header.js` in the root of your project, this header will be used verbatim
- * as the ${header} template variable instead of the generic header template, which is populated with 
- * information from Manifest.json
- * 
- * (Contrib support is not yet implemented)
- * 
+ * Handles contrib libraries
  */
 qx.Class.define("qxcli.commands.Add", {
   extend: qxcli.commands.Command,
+  
   statics: {
+    
     getYargsCommand: function() {
       return {
-        command: 'add <template> <classname> [options]',
-        describe: 'adds a new class file to the current project, based on a template.',
-        builder: {
-          "extend":{
-            alias : "e",
-            describe: 'the class to extend'
-          },
-          "import":{
-            describe: 'import the template to the `templates/class` folder of the current project, where it can be customized'
-          },
-          "force":{
-            alias : "f",
-            describe: 'overwrite an existing file'
-          },                        
-          // "verbose":{
-          //   alias : "v",
-          //   describe: 'Verbose logging'
-          // }                    
-        },
-        handler: function(argv) {
-          return new qxcli.commands.Add(argv)
-            .process()
-            .catch((e) => {
-              console.error(e.stack || e.message);
-              process.exit(1);
-            });
+        command : "add <command> [options]",
+        desc : "adds new elements to an existing qooxdoo application/library",
+        builder : function (yargs) { 
+          return yargs 
+            .commandDir('../../../commands/add_commands') 
+            .demandCommand() 
+            .showHelpOnFail() 
+        }, 
+        handler : function(argv){
         }
-      };
+      }
     }
-  },
-
+  }, 
+  
   members: {
-    process: async function() {
-      let argv = this.argv;
-
-      // read Manifest.json
-      let manifestPath = path.join( process.cwd(), "Manifest.json" );
-      if( ! await fs.existsAsync( manifestPath )){
-        throw new qxcli.Utils.UserError("No Manifest.json file in this directory. Please go to the your project root.");
-      }
-      let manifestJson = await this.parseJsonFile(manifestPath);
-      
-      // prepare template vars
-      let values = Object.assign({}, manifestJson.info, manifestJson.provides );
-      // @todo Add support for authors, ask interactively if author info should be taken
-      // from Manifest or entered manually, then create string representation to insert.
-      values.authors = "";
-      values.classname = argv.classname;
-      values.extend = argv.extend ? argv.extend : "qx.core.Object";
-      
-      // @todo ask interactively for copyright holder, create a setting in Manifest.json
-      values.copyright = (new Date).getFullYear();
-
-      // check top-level namespace
-      let class_namespaces = argv.classname.split(/\./);
-      if( class_namespaces[0] !== values.namespace ){
-        throw new qxcli.Utils.UserError(`Invalid top namespace '${class_namespaces[0]}'. Must be '${values.namespace}'.`);
-      }   
-
-      // get path to the template file
-      let template_name = argv.template;
-      let template_path;
-      let potential_dirs = [
-        // 1. in the templates/class dir of the current project
-        path.join( process.cwd(), "templates"),
-        // 2. in the templates/class dir of cli
-        this.getTemplateDir()
-        // 3. @todo: in a contrib library's templates dir
-      ];
-      let found = false;
-      for( let dir of potential_dirs ){
-        template_path = path.join( dir, "class", template_name + ".tmpl.js");
-        if ( await fs.existsAsync( template_path ) ){
-          found = true; break; 
-        }
-      }
-      if( ! found ){
-        throw new qxcli.Utils.UserError(`Template ${template_name} does not exist.`);
-      }
-      let template = await fs.readFileAsync(template_path,"utf-8");
-
-      // handle header macro in the project root
-      let header_template;
-      let header_template_path = path.join( process.cwd(), "header.js" );
-      try{
-        header_template = fs.readFileSync(header_template_path,"utf-8");
-      } catch(e) {
-        // if none exists, use header template in the same folder as the template itself
-        header_template_path =  path.join( path.dirname(template_path), "header.tmpl.js" );
-        try {
-          header_template = fs.readFileSync(header_template_path,"utf-8");
-        } catch(e){}
-      }
-      if ( header_template ){
-        // replace template vars in header
-        if( header_template_path.includes(".tmpl.js")) {
-          for( let var_name in values ){
-            header_template = header_template.replace(new RegExp(`\\$\{${var_name}\}`,"g"), values[var_name] );
-          }
-        }
-        values.header = header_template;
-      }
-      
-      // replace template vars
-      let final_content = template;
-      for( let var_name in values ){
-        final_content = final_content.replace(new RegExp(`\\$\{${var_name}\}`,"g"), values[var_name] );
-      }
-      // check if file already exists
-      let relative_path = path.join( "source", "class", ...class_namespaces ) + '.js';
-      let absolute_path = path.join( process.cwd(), relative_path );
-      let file_exists = false;
-      try{
-        fs.accessSync(absolute_path);
-        file_exists = true;
-      } catch(e){}
-      if( file_exists && ! argv.force){
-        throw new qxcli.Utils.UserError(`Class file ${relative_path} already exists. Use --force to overwrite it` );
-      } 
-      
-
-      // write out new class file
-      try {
-        require('mkdirp').sync( path.dirname( absolute_path ), 0o755 );
-        await fs.writeFileAsync( absolute_path, final_content, "utf-8")
-      } catch(e){
-        throw new qxcli.Utils.UserError(`Cannot write to ${absolute_path}: ${e.message}` );
-      }
-
-      // import
-      if( argv.import ){
-        let local_templates_path = path.join( process.cwd(), "templates", "class" );
-        let local_copy_path = path.join( local_templates_path, path.basename( template_path ) );
-        try {
-          require('mkdirp').sync( local_templates_path, 0o755 );
-          await fs.writeFileAsync( local_copy_path, template, "utf-8" );
-        } catch(e) {
-          throw new qxcli.Utils.UserError(`Cannot copy template to ${local_templates_path}: ${e.message}` );
-        }
-      }      
-    }
+    // place for common methods.
   }
 });
