+++ templates/class/default.tmpl.js
